[いいね機能]非同期通信化の実装

+++ app/controllers/public/favorites_controller.rb
@@ -1,15 +1,15 @@
 class Public::FavoritesController < ApplicationController
+  before_action :authenticate_user!
+
   def create
-    practice_post = PracticePost.find(params[:practice_post_id])
-    favorite = current_user.favorites.new(practice_post_id: practice_post.id)
+    @practice_post = PracticePost.find(params[:practice_post_id])
+    favorite = current_user.favorites.new(practice_post_id: @practice_post.id)
     favorite.save
-    redirect_to request.referer
   end
 
   def destroy
-    practice_post = PracticePost.find(params[:practice_post_id])
-    favorite = current_user.favorites.find_by(practice_post_id: practice_post.id)
-    favorite.destroy
-    redirect_to request.referer
+    @practice_post = PracticePost.find(params[:practice_post_id])
+    favorite = current_user.favorites.find_by(practice_post_id: @practice_post.id)
+    favorite.destroy if favorite.present?
   end
 end
